fix: incorrect types for Accuracy and Bearing in PostDeathDto and PostKillDto

+++ AirsoftBmsApp/Model/Dto/Death/PostDeathDto.cs
@@ -4,7 +4,7 @@ public class PostDeathDto
 {
     public double Longitude { get; set; }
     public double Latitude { get; set; }
-    public double Accuracy { get; set; }
-    public double Bearing { get; set; }
+    public short Accuracy { get; set; }
+    public short Bearing { get; set; }
     public DateTimeOffset Time { get; set; }
 }
